Richer responses from readAll*
If Buffer.readFrom* exited because their rid was (or became)
unavailable, return { closed: true, content: buffer }.  If it exited
cleanly, return { closed: false, content: buffer }.

If readFrom exited because of another error, readFrom returns
{ aborted: that_error, content: buffer }.  (Need this for handling later
refinements to Process. Not sure whether readFromSync should do this
too; for now that just propagates the error.)

In all cases, the buffer read so far is provided.