Add mode to make_temp
Add `mode` option for makeTempDir and makeTempFile. As with mkdir, this
argument is only used when a new dir or file is created. Defaults to
0o700 for directories and 0o600 for files.

We make the tests verify that the `mode` used is subject to the
process's umask (here just assumed to be the default 0o022 on Unix, or 0
on Windows).