writeFile: createNew option, better mode tests
This introduces a change of functionality. To make the operation more
atomic, and conform better to other parts of the API, writeFile will now
only apply `mode` if it creates a new file, not if it overwrites or
appends to an existing file.

Previously, writeFile did a chmod on the target file no matter what. If
that's the behavior a user wants, they can do it easily enough with
chmod. (And it never was atomic.)