Change numeric parameters for op_seek
The whence argument in op_seek was received as i32 and cast to u32.
Can't see any good reason for the recast?

The offset parameter to op_seek was i32, which we were casting to u64
(so using it as a u32) for SEEK_START, and extending to i64 for the
other seek modes.

Instead made offset i64, requiring it to be non-negative (so effectively
u63) for SEEK_START, permitting to be negative for other seek modes.
That's what libc also seems to be doing: it uses off_t, which is usually
i64.