copyFile: create[New], clobber
Add create and createNew (and clobber) options for copyFile. These work
as for writeFile. The fastest implementation is used when { create:
true, createNew: false } (this is the default, and matches the older
implementation).

To implement createNew atomically, we have to fall back to what I think
is a slower implementation, where the data gets copied from kernel space
to user space, and back.