Better mode tests for stat/writeFile
We verify that the `mode` applied by writeFile are only used when
creating a new file; also that they are subject to the process's umask
(here just assumed to be the default 0o022 on Unix, or 0 on Windows).

Perhaps this test should be moved from stat_test.ts to
write_file_test.ts?