Require trunc len to be u63
The len parameter to truncate was u64. Odd to use a bigger type for
truncate than for seek (where we were effectively letting offset be
u32); and indeed the underlying libc expects an off_t, which is i64.

Instead made this variable i64, requiring it to be non-negative (so
effectively u63). This matches the treatment of offset for op_seek when
whence is SEEK_START.