ops/fs.rs: readdir: build Vec by hand
When we shift from std::fs::read_dir to the tokio::fs version (#4188), the processing of entries will be an async function, and Rust doesn't currently have a stable implementation of async closures. So instead of using `filter_map` with a closure, we iterate and build the Vec by hand. This makes the shift to tokio::fs more straightforward.