Add clobber option for open
This is the first in a series of commits adding a `clobber` option to
filesystem operations. The term `clobber` comes from the shell's
`noclobber` option.

Mostly, `clobber: false` is just another way to write `createNew: true`.
There are some cases though where this functionality could be useful but
it's unnatural to use `createNew` to express it. For example,
`rename(oldpath, newpath, { clobber: false })`
will fail if there's already a file at `newpath`, instead of the
current (default) behavior of overwriting newpath.